refactor: move the brand permissions to a dedicated file
